# Generator performance fixes