--- a/hack/crossplane/apis/cache/v1beta20200601/redis_types_gen.go
+++ b/hack/crossplane/apis/cache/v1beta20200601/redis_types_gen.go
@@ -325,6 +325,8 @@ const (
 )
 
 type RedisProperties_Status_RedisConfiguration struct {
+	AdditionalProperties map[string]string `json:"additionalProperties,omitempty"`
+
 	// AofStorageConnectionString0: First storage account connection string
 	AofStorageConnectionString0 *string `json:"aof-storage-connection-string-0,omitempty"`
 
@@ -356,8 +358,7 @@ type RedisProperties_Status_RedisConfiguration struct {
 	RdbBackupMaxSnapshotCount *string `json:"rdb-backup-max-snapshot-count,omitempty"`
 
 	// RdbStorageConnectionString: The storage account connection string for storing rdb file
-	RdbStorageConnectionString *string           `json:"rdb-storage-connection-string,omitempty"`
-	additionalProperties       map[string]string `json:"additionalProperties,omitempty"`
+	RdbStorageConnectionString *string `json:"rdb-storage-connection-string,omitempty"`
 }
 
 // Generated from: https://schema.management.azure.com/schemas/2020-06-01/Microsoft.Cache.json#/definitions/Sku