--- a/v2/api/authorization/v1beta20200801preview/role_assignment_types_gen.go
+++ b/v2/api/authorization/v1beta20200801preview/role_assignment_types_gen.go
@@ -804,7 +804,7 @@ func (assignments *RoleAssignments_Spec) ConvertToARM(resolved genruntime.Conver
 
 	// Set property ‘Tags’:
 	if assignments.Tags != nil {
-		result.Tags = make(map[string]string)
+		result.Tags = make(map[string]string, len(assignments.Tags))
 		for key, value := range assignments.Tags {
 			result.Tags[key] = value
 		}
@@ -894,7 +894,7 @@ func (assignments *RoleAssignments_Spec) PopulateFromARM(owner genruntime.Arbitr
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		assignments.Tags = make(map[string]string)
+		assignments.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			assignments.Tags[key] = value
 		}