--- a/v2/api/cache/v1alpha1api20201201/redis_firewall_rule_types_gen.go
+++ b/v2/api/cache/v1alpha1api20201201/redis_firewall_rule_types_gen.go
@@ -560,7 +560,7 @@ func (rules *RedisFirewallRules_Spec) ConvertToARM(resolved genruntime.ConvertTo
 
 	// Set property ‘Tags’:
 	if rules.Tags != nil {
-		result.Tags = make(map[string]string)
+		result.Tags = make(map[string]string, len(rules.Tags))
 		for key, value := range rules.Tags {
 			result.Tags[key] = value
 		}
@@ -614,7 +614,7 @@ func (rules *RedisFirewallRules_Spec) PopulateFromARM(owner genruntime.Arbitrary
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		rules.Tags = make(map[string]string)
+		rules.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			rules.Tags[key] = value
 		}