--- a/v2/api/cache/v1alpha1api20201201/redis_linked_server_types_gen.go
+++ b/v2/api/cache/v1alpha1api20201201/redis_linked_server_types_gen.go
@@ -615,7 +615,7 @@ func (servers *RedisLinkedServers_Spec) ConvertToARM(resolved genruntime.Convert
 
 	// Set property ‘Tags’:
 	if servers.Tags != nil {
-		result.Tags = make(map[string]string)
+		result.Tags = make(map[string]string, len(servers.Tags))
 		for key, value := range servers.Tags {
 			result.Tags[key] = value
 		}
@@ -671,7 +671,7 @@ func (servers *RedisLinkedServers_Spec) PopulateFromARM(owner genruntime.Arbitra
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		servers.Tags = make(map[string]string)
+		servers.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			servers.Tags[key] = value
 		}