--- a/v2/api/cache/v1alpha1api20201201/redis_types_gen.go
+++ b/v2/api/cache/v1alpha1api20201201/redis_types_gen.go
@@ -571,7 +571,7 @@ func (resource *RedisResource_Status) PopulateFromARM(owner genruntime.Arbitrary
 	// copying flattened property:
 	if typedInput.Properties != nil {
 		if typedInput.Properties.RedisConfiguration != nil {
-			resource.RedisConfiguration = make(map[string]string)
+			resource.RedisConfiguration = make(map[string]string, len(typedInput.Properties.RedisConfiguration))
 			for key, value := range typedInput.Properties.RedisConfiguration {
 				resource.RedisConfiguration[key] = value
 			}
@@ -657,7 +657,7 @@ func (resource *RedisResource_Status) PopulateFromARM(owner genruntime.Arbitrary
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		resource.Tags = make(map[string]string)
+		resource.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			resource.Tags[key] = value
 		}
@@ -667,7 +667,7 @@ func (resource *RedisResource_Status) PopulateFromARM(owner genruntime.Arbitrary
 	// copying flattened property:
 	if typedInput.Properties != nil {
 		if typedInput.Properties.TenantSettings != nil {
-			resource.TenantSettings = make(map[string]string)
+			resource.TenantSettings = make(map[string]string, len(typedInput.Properties.TenantSettings))
 			for key, value := range typedInput.Properties.TenantSettings {
 				resource.TenantSettings[key] = value
 			}
@@ -1097,7 +1097,7 @@ func (redis *Redis_Spec) ConvertToARM(resolved genruntime.ConvertToARMResolvedDe
 		result.Properties.PublicNetworkAccess = &publicNetworkAccess
 	}
 	if redis.RedisConfiguration != nil {
-		result.Properties.RedisConfiguration = make(map[string]string)
+		result.Properties.RedisConfiguration = make(map[string]string, len(redis.RedisConfiguration))
 		for key, value := range redis.RedisConfiguration {
 			result.Properties.RedisConfiguration[key] = value
 		}
@@ -1139,7 +1139,7 @@ func (redis *Redis_Spec) ConvertToARM(resolved genruntime.ConvertToARMResolvedDe
 		result.Properties.SubnetId = &subnetId
 	}
 	if redis.TenantSettings != nil {
-		result.Properties.TenantSettings = make(map[string]string)
+		result.Properties.TenantSettings = make(map[string]string, len(redis.TenantSettings))
 		for key, value := range redis.TenantSettings {
 			result.Properties.TenantSettings[key] = value
 		}
@@ -1147,7 +1147,7 @@ func (redis *Redis_Spec) ConvertToARM(resolved genruntime.ConvertToARMResolvedDe
 
 	// Set property ‘Tags’:
 	if redis.Tags != nil {
-		result.Tags = make(map[string]string)
+		result.Tags = make(map[string]string, len(redis.Tags))
 		for key, value := range redis.Tags {
 			result.Tags[key] = value
 		}
@@ -1219,7 +1219,7 @@ func (redis *Redis_Spec) PopulateFromARM(owner genruntime.ArbitraryOwnerReferenc
 	// copying flattened property:
 	if typedInput.Properties != nil {
 		if typedInput.Properties.RedisConfiguration != nil {
-			redis.RedisConfiguration = make(map[string]string)
+			redis.RedisConfiguration = make(map[string]string, len(typedInput.Properties.RedisConfiguration))
 			for key, value := range typedInput.Properties.RedisConfiguration {
 				redis.RedisConfiguration[key] = value
 			}
@@ -1289,7 +1289,7 @@ func (redis *Redis_Spec) PopulateFromARM(owner genruntime.ArbitraryOwnerReferenc
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		redis.Tags = make(map[string]string)
+		redis.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			redis.Tags[key] = value
 		}
@@ -1299,7 +1299,7 @@ func (redis *Redis_Spec) PopulateFromARM(owner genruntime.ArbitraryOwnerReferenc
 	// copying flattened property:
 	if typedInput.Properties != nil {
 		if typedInput.Properties.TenantSettings != nil {
-			redis.TenantSettings = make(map[string]string)
+			redis.TenantSettings = make(map[string]string, len(typedInput.Properties.TenantSettings))
 			for key, value := range typedInput.Properties.TenantSettings {
 				redis.TenantSettings[key] = value
 			}