--- a/v2/api/cache/v1beta20210301/redis_enterprise_database_types_gen.go
+++ b/v2/api/cache/v1beta20210301/redis_enterprise_database_types_gen.go
@@ -817,7 +817,7 @@ func (databases *RedisEnterpriseDatabases_Spec) ConvertToARM(resolved genruntime
 
 	// Set property ‘Tags’:
 	if databases.Tags != nil {
-		result.Tags = make(map[string]string)
+		result.Tags = make(map[string]string, len(databases.Tags))
 		for key, value := range databases.Tags {
 			result.Tags[key] = value
 		}
@@ -916,7 +916,7 @@ func (databases *RedisEnterpriseDatabases_Spec) PopulateFromARM(owner genruntime
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		databases.Tags = make(map[string]string)
+		databases.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			databases.Tags[key] = value
 		}