--- a/v2/api/compute/v1alpha1api20210701/image_types_gen.go
+++ b/v2/api/compute/v1alpha1api20210701/image_types_gen.go
@@ -496,7 +496,7 @@ func (image *Image_Status) PopulateFromARM(owner genruntime.ArbitraryOwnerRefere
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		image.Tags = make(map[string]string)
+		image.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			image.Tags[key] = value
 		}
@@ -740,7 +740,7 @@ func (images *Images_Spec) ConvertToARM(resolved genruntime.ConvertToARMResolved
 
 	// Set property ‘Tags’:
 	if images.Tags != nil {
-		result.Tags = make(map[string]string)
+		result.Tags = make(map[string]string, len(images.Tags))
 		for key, value := range images.Tags {
 			result.Tags[key] = value
 		}
@@ -824,7 +824,7 @@ func (images *Images_Spec) PopulateFromARM(owner genruntime.ArbitraryOwnerRefere
 
 	// Set property ‘Tags’:
 	if typedInput.Tags != nil {
-		images.Tags = make(map[string]string)
+		images.Tags = make(map[string]string, len(typedInput.Tags))
 		for key, value := range typedInput.Tags {
 			images.Tags[key] = value
 		}