… .golangci.yml
… hack/crossplane/apis/cache/v1beta20200601/zz_generated.deepcopy.go
… v2/api/compute/v1alpha1api20201201/virtual_machine_scale_set_types_gen.go
… v2/api/compute/v1alpha1api20201201/virtual_machine_types_gen.go
… v2/api/compute/v1beta20200930/disk_types_gen.go
… v2/api/compute/v1beta20200930/snapshot_types_gen.go
… v2/api/compute/v1beta20201201/virtual_machine_scale_set_types_gen.go
… v2/api/compute/v1beta20201201/virtual_machine_types_gen.go
… v2/api/compute/v1beta20210701/image_types_gen.go
… v2/api/containerregistry/v1alpha1api20210901/registry_types_gen.go
… v2/api/containerregistry/v1beta20210901/registry_types_gen.go
… v2/api/containerservice/v1alpha1api20210501/managed_cluster_types_gen.go
… v2/api/containerservice/v1alpha1api20210501/managed_clusters_agent_pool_types_gen.go
… v2/api/containerservice/v1beta20210501/managed_cluster_types_gen.go
… v2/api/containerservice/v1beta20210501/managed_clusters_agent_pool_types_gen.go
… v2/api/dbformariadb/v1beta20180601/database_types_gen.go
… v2/api/dbformariadb/v1beta20180601/server_types_gen.go
… v2/api/dbformysql/v1alpha1api20210501/flexible_server_types_gen.go
… v2/api/dbformysql/v1alpha1api20210501/flexible_servers_database_types_gen.go
… v2/api/dbformysql/v1alpha1api20210501/flexible_servers_firewall_rule_types_gen.go
… v2/api/dbformysql/v1beta20210501/flexible_server_types_gen.go
… v2/api/dbformysql/v1beta20210501/flexible_servers_database_types_gen.go
… v2/api/dbformysql/v1beta20210501/flexible_servers_firewall_rule_types_gen.go
… v2/api/dbforpostgresql/v1alpha1api20210601/flexible_server_types_gen.go
… v2/api/dbforpostgresql/v1alpha1api20210601/flexible_servers_configuration_types_gen.go
… v2/api/dbforpostgresql/v1alpha1api20210601/flexible_servers_database_types_gen.go
… v2/api/dbforpostgresql/v1alpha1api20210601/flexible_servers_firewall_rule_types_gen.go
… v2/api/dbforpostgresql/v1beta20210601/flexible_server_types_gen.go
… v2/api/dbforpostgresql/v1beta20210601/flexible_servers_configuration_types_gen.go
… v2/api/dbforpostgresql/v1beta20210601/flexible_servers_database_types_gen.go