We've had a bit of a regression recently in generator performance, which leads to increased time spent doing the edit-test loop.

Here's a grab-bag of fixes driven by `pprof` analysis.

Note that a lot of time is being spent in GC due to the number of object copies we make, so much of the changes below are about reducing the number of allocations.

# Summary

Time spent in memory/GC improved a lot (cumulative times below):

- malloc from 4.59s to 2.63s
- gcDrain from 18.91s to 10.73s

Overall time spent in TypeVisitor (one of the main hotspots) from 11.56s to 6.55s.

Biggest win in terms of allocations is that `ObjectType.copy` used to allocate >2GB of data and now allocates <150MB.

The main thing left to do would be somehow speeding up `TypeDefinitionSet`, as there is a lot of time spent in hashing `TypeNames` and copying maps and so on. A nice fix would be to replace `TypeName` with some lightweight standin containing just an integer value. 

JSON loading also takes a significant part of the runtime but there's not a lot we can do about that.

# Changes made

## Add short-circuiting to `MakeIdentityVisitOfObjectType`, and `WithProperties` on `ObjectType`. 

Improves runtimes of type visitors:
| Before | After |
|-|-|
| ![image](https://user-images.githubusercontent.com/12575/170151562-f566bf7a-78da-419d-9c69-0c961d7e1354.png) | ![image](https://user-images.githubusercontent.com/12575/170151641-d77f869c-259b-4470-8a60-91972bc3862f.png) |

Memory/GC churn is also greatly reduced:
| Before | After |
|-|-|
| ![image](https://user-images.githubusercontent.com/12575/170152759-75b9e1ef-e7c7-4fa4-92a9-7400421e7572.png) GC | ![image](https://user-images.githubusercontent.com/12575/170152788-8601dff9-c868-4a1a-b2ec-a4c0ffed0d8d.png) |

This reduces runtime by quite a bit for pipeline stages that visit lots of types, for example:
- `Remove validation from all status properties, completed in 2.796s`
- `Remove validation from all status properties, completed in 1.887s`

## Right-size allocations

Enable the `prealloc` linter and follow its suggestions. Also right-size various maps (not detected by `prealloc`) and the map allocations in generated code.

## Reduce memory usage of `TypeMatcher`

`TypeMatcher` was holding onto 14 MB of strings. Now it holds onto none, since it only needs to track `count > 0`.

This reduced snapshot memory usage from 77MB to 59MB.

![image](https://user-images.githubusercontent.com/12575/170165128-43599ef7-ee57-4868-9be2-85587c0af6be.png)

## Make `LocalPackageReference.Version` non-allocating

This was responsible for a lot of garbage:
![image](https://user-images.githubusercontent.com/12575/170368966-e29b92de-eaf0-4576-a559-5bcd5cbe875c.png)


## Make `MakeIdentityVisitOfObjectType` not invoke `PropertySet.AsSlice`

There's no need for this as the function does not require the properties to be in sorted order. Removes all these allocations:

![image](https://user-images.githubusercontent.com/12575/170370236-177238ff-0471-47d3-95a4-e9b3f0723f3b.png)

(NB: this is improved again by a later change, below, to remove `PropertySet` copies.)

## Shortcut for `IdentityVisitOfTypeName`

When we are using a `TypeVisitor` to visit a `TypeName` and the visitor for `TypeName`s is `IdentityVisitOfTypeName`, skip calling the visitor and return the `Type` directly. This avoids reboxing the `TypeName` into a new `Type`.

This removes most of these allocations:
![image](https://user-images.githubusercontent.com/12575/170372824-bf3c4eab-3ad7-4b38-a615-ce50780184c0.png)

## Introduce `ReadOnlyPropertySet` interface to reduce defensive copies

This only allows readonly operations on the property set. A `PropertySet` can be obtained (to retain preëxisting behaviour) by invoking `Copy()`.

```go
type ReadOnlyPropertySet interface {
	AsSlice() []*PropertyDefinition
	ContainsProperty(name PropertyName) bool
	Get(name PropertyName) (*PropertyDefinition, bool)
	Find(predicate func(*PropertyDefinition) bool) (*PropertyDefinition, bool)
	Copy() PropertySet
	Equals(other ReadOnlyPropertySet, overrides EqualityOverrides) bool
	First() *PropertyDefinition
	ForEach(func(def *PropertyDefinition))
	IsEmpty() bool
	Len() int
}
```


This reduces the number of property copies by a lot:

| Before | After |
|-|-|
| ![MicrosoftTeams-image (5)](https://user-images.githubusercontent.com/12575/170379099-def73947-b05e-4ebc-9e35-fa4c0c35ff48.png) | ![image](https://user-images.githubusercontent.com/12575/170379222-2736feea-36db-4bdf-92b0-bef6adbc0a02.png)  | 

## Make `ObjectType` use readonly maps internally to reduce defensive copies

This is the logical conclusion of the above. By using readonly maps internally, we only make copies when needed.

Total allocations from maps decreases a lot:
| Before | After |
|-|-|
| ![image](https://user-images.githubusercontent.com/12575/170406701-c2bcd159-21b0-4d5b-bf12-f1cbc566d93b.png) | ![image](https://user-images.githubusercontent.com/12575/170406906-ee1f3eb3-d895-4746-8cd6-015c1d355e9d.png) |

## Make `PropertyDefinition` use readonly maps internally

Same idea as previous, and the last change.

| Before | After |
|-|-|
| ![image](https://user-images.githubusercontent.com/12575/170409524-42b00291-2b89-4c51-866c-10d6dec8dc93.png) | ![image](https://user-images.githubusercontent.com/12575/170409571-ef089ca9-ce60-4930-97d2-fbd83ab8b1fd.png) |