@Porges 2022-05-26
hmm, one ordering/nondeterminism issue to chase down before merge